exchange implements and other queue

+++ GeekShopping/GeekShopping.OrderAPI/Repositories/Interfaces/IOrderRepository.cs
@@ -5,6 +5,6 @@ namespace GeekShopping.OrderAPI.Repositories.Interfaces
     public interface IOrderRepository
     {
         Task<bool> AddOrder(OrderHeader header);
-        Task UpdateOrderPaymentStatus(int orderHeaderId, bool status);
+        Task UpdateOrderPaymentStatus(long orderHeaderId, bool status);
     }
 }
